+++ src/header/GameEnvironment/CandyGameState.hpp
@@ -20,7 +20,7 @@ public:
     CandyGameState();
 
     // Specified ctor
-    CandyGameState(const vector<CandyPlayer>& playerList, int roundNum, unsigned long long points, const TableCard& tableCard, 
+    CandyGameState(const vector<CandyPlayer>& playerList, int roundNum, uint64_t points, const TableCard& tableCard, 
         const GameDeckCard& deckCard, const AbilityDeckCard& abilities, int reversePlayerId);
 
     // cctor
@@ -43,8 +43,8 @@ public:
     
     AbilityDeckCard& getAbilities();
 
-    static const unsigned long long initialPoint;
-    static const unsigned long long winnerPoint;
+    static const uint64_t initialPoint;
+    static const uint64_t winnerPoint;
 };
 
 #endif
